Use `[data-turbo-busy]` everywhere
Toggle the `[data-turbo-busy]` attribute to be present on a
`<turbo-frame>` element when its navigating, a `<form>` element when
it's submitting, or on the `<html>` element during any request, then
remove the attribute after the submission or navigation completes.

This is a change from the `turbo-frame[busy]` attribute, but aims to
achieve the same purpose.

By unifying a single, consistent attribute, consumer applications can
use a single attribute CSS selector at different depths within their
page to hide or show loading indicators.

For example, consider a loading indication spinner animation within a
submit button that animates while a submission is in progress:

```html
<style>
  .loading-spinner                   { display: none; }
  [data-turbo-busy] .loading-spinner { display: block; }
</style>

<form action="/posts">
  <!-- ... -->
  <button type="submit">
    <span class="loading-spinner>...</span>
    <span>Create Post</span>
  </button>
</form>
```

Similarly, consider a `turbo-frame` scoped loading bar:

```html
<style>
  .loading-bar                   { display: none; }
  [data-turbo-busy] .loading-bar { display: block; }
</style>
<a href="/content" data-turbo-frame="content">...</a>
<turbo-frame id="content">
  <progress id="content-loading-bar" class="loading-bar" value="0" data-turbo-permanent></progress>

  <!-- ... -->
</turbo-frame>
```

Finally, consider the same `.loading-bar` for page-wide navigation:

```html
<style>
  .loading-bar                   { display: none; }
  [data-turbo-busy] .loading-bar { display: block; }
</style>

<nav>
  <progress id="loading-bar" class="loading-bar" value="0" data-turbo-permanet></progress>
</nav>

<!-- ... -->

<a href="/posts/1">...</a>
```