Replace LinkInterceptor with LinkClickObserver
Follow-up to https://github.com/hotwired/turbo/pull/382

---

In the same style as [hotwired/turbo#382][], replace instances of
`LinkInterceptor` with `LinkClickObserver`, making the necessary
interface changes in classes that used to extend
`LinkInterceptorDelegate`.

Conditional logic that was once covered in the `LinkInterceptor` is
moved into the predicate methods of the delegates (for example, the
`FrameController.willFollowLinkToLocation` and
`FrameRedirector.willFollowLinkToLocation`).

[hotwired/turbo#382]: https://github.com/hotwired/turbo/pull/382